🚚 message の中身を messages.go に移動
